+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - `set_module_conf`: https://docs.thingsdb.net/v0/thingsdb-api/set_module_conf/
   - `set_module_scope`: https://docs.thingsdb.net/v0/thingsdb-api/set_module_scope/
   - `restart_module`: https://docs.thingsdb.net/v0/node-api/restart_module/
+* Added `equals` function for comparing things, issue #144.
 
 # v0.10.1